+++ backend-rust/src/indexer.rs
@@ -618,10 +618,16 @@ impl ProcessEvent for BlockProcessor {
         PreparedBlock::batch_save(batch, &mut new_context, &mut tx).await?;
         for block in batch {
             for item in block.prepared_block_items.iter() {
-                item.save(&mut tx).await?;
+                item.save(&mut tx).await.with_context(|| {
+                    format!(
+                        "Failed processing block item with hash {} for block height {}",
+                        item.block_item_hash, item.block_height
+                    )
+                })?;
             }
             block.special_transaction_outcomes.save(&mut tx).await?;
             block.baker_unmark_suspended.save(&mut tx).await?;
+
             out.push_str(format!("\n- {}:{}", block.height, block.hash).as_str());
         }
         process_release_schedules(new_context.last_block_slot_time, &mut tx)
@@ -872,6 +878,7 @@ impl PreparedBlock {
         )
         .await?;
         let baker_unmark_suspended = PreparedUnmarkPrimedForSuspension::prepare(data)?;
+
         Ok(Self {
             hash,
             height,
@@ -934,18 +941,26 @@ impl PreparedBlock {
         }
 
         sqlx::query!(
-            "INSERT INTO blocks
-  (height, hash, slot_time, block_time, baker_id, total_amount, total_staked, cumulative_num_txs)
-SELECT * FROM UNNEST(
-  $1::BIGINT[],
-  $2::TEXT[],
-  $3::TIMESTAMPTZ[],
-  $4::BIGINT[],
-  $5::BIGINT[],
-  $6::BIGINT[],
-  $7::BIGINT[],
-  $8::BIGINT[]
-);",
+            "INSERT INTO blocks (
+                height, 
+                hash, 
+                slot_time, 
+                block_time, 
+                baker_id, 
+                total_amount, 
+                total_staked, 
+                cumulative_num_txs
+            )
+            SELECT * FROM UNNEST(
+                $1::BIGINT[],
+                $2::TEXT[],
+                $3::TIMESTAMPTZ[],
+                $4::BIGINT[],
+                $5::BIGINT[],
+                $6::BIGINT[],
+                $7::BIGINT[],
+                $8::BIGINT[]
+            );",
             &heights,
             &hashes,
             &slot_times,
@@ -1031,6 +1046,7 @@ SELECT * FROM UNNEST(
 /// Database operation for adding new row into the account statement table.
 /// This reads the current balance of the account and assumes the balance is
 /// already updated with the amount part of the statement.
+#[derive(Debug)]
 struct PreparedAccountStatement {
     canonical_address: CanonicalAccountAddress,
     amount:            i64,
@@ -1075,13 +1091,14 @@ impl PreparedAccountStatement {
         .execute(tx.as_mut())
         .await?
         .ensure_affected_one_row()
-        .context("Failed insert into account_statements")?;
+        .with_context(|| format!("Failed insert into account_statements: {:?}", self))?;
 
         Ok(())
     }
 }
 
 /// Prepared block item (transaction), ready to be inserted in the database
+#[derive(Debug)]
 struct PreparedBlockItem {
     /// Hash of the transaction
     block_item_hash:   String,
@@ -1254,7 +1271,8 @@ impl PreparedBlockItem {
             reject
         )
         .fetch_one(tx.as_mut())
-        .await?;
+        .await
+        .context("Failed inserting into transactions")?;
         // Note that this does not include account creation. We handle that when saving
         // the account creation event.
         sqlx::query!(
@@ -1281,12 +1299,18 @@ impl PreparedBlockItem {
         .ensure_affected_rows(self.affected_accounts.len().try_into()?)
         .context("Failed incrementing num_txs for account")?;
 
-        self.prepared_event.save(tx, tx_idx).await?;
+        self.prepared_event.save(tx, tx_idx).await.with_context(|| {
+            format!(
+                "Failed processing block item event from {:?} transaction",
+                self.transaction_type
+            )
+        })?;
         Ok(())
     }
 }
 
 /// Different types of block item events that can be prepared.
+#[derive(Debug)]
 enum PreparedBlockItemEvent {
     /// A new account got created.
     AccountCreation(PreparedAccountCreation),
@@ -1351,6 +1375,7 @@ impl PreparedBlockItemEvent {
     }
 }
 
+#[derive(Debug)]
 struct PreparedAccountTransaction {
     /// Update the balance of the sender account with the cost (transaction
     /// fee).
@@ -1359,6 +1384,7 @@ struct PreparedAccountTransaction {
     event: PreparedEventEnvelope,
 }
 
+#[derive(Debug)]
 enum PreparedEvent {
     /// A transfer of CCD from one account to another account.
     CcdTransfer(PreparedCcdTransferEvent),
@@ -1651,17 +1677,42 @@ impl PreparedEvent {
         protocol_version: ProtocolVersion,
     ) -> anyhow::Result<()> {
         match self {
-            PreparedEvent::CcdTransfer(event) => event.save(tx, tx_idx).await,
-            PreparedEvent::EncryptedBalance(event) => event.save(tx, tx_idx).await,
-            PreparedEvent::BakerEvents(event) => event.save(tx, tx_idx, protocol_version).await,
-            PreparedEvent::ModuleDeployed(event) => event.save(tx, tx_idx).await,
-            PreparedEvent::ContractInitialized(event) => event.save(tx, tx_idx).await,
-            PreparedEvent::ContractUpdate(event) => event.save(tx, tx_idx).await,
-            PreparedEvent::AccountDelegationEvents(event) => {
-                event.save(tx, tx_idx, protocol_version).await
-            }
-            PreparedEvent::ScheduledTransfer(event) => event.save(tx, tx_idx).await,
-            PreparedEvent::RejectedTransaction(event) => event.save(tx, tx_idx).await,
+            PreparedEvent::CcdTransfer(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event of CCD transfer"),
+            PreparedEvent::EncryptedBalance(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event of encrypted balance"),
+            PreparedEvent::BakerEvents(event) => event
+                .save(tx, tx_idx, protocol_version)
+                .await
+                .context("Failed processing block item event with baker event"),
+            PreparedEvent::ModuleDeployed(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event with module deploy"),
+            PreparedEvent::ContractInitialized(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event with contract initialized"),
+            PreparedEvent::ContractUpdate(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event with contract update"),
+            PreparedEvent::AccountDelegationEvents(event) => event
+                .save(tx, tx_idx, protocol_version)
+                .await
+                .context("Failed processing block item event with account delegation event"),
+            PreparedEvent::ScheduledTransfer(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event with scheduled transfer"),
+            PreparedEvent::RejectedTransaction(event) => event
+                .save(tx, tx_idx)
+                .await
+                .context("Failed processing block item event with rejected event"),
             PreparedEvent::NoOperation => Ok(()),
         }
     }
@@ -1669,6 +1720,7 @@ impl PreparedEvent {
 
 /// Contains metadata required for event processing that is not directly tied to
 /// individual events.
+#[derive(Debug)]
 struct EventMetadata {
     protocol_version: ProtocolVersion,
 }
@@ -1679,6 +1731,7 @@ struct EventMetadata {
 /// during which other baker-related transactions could still occur, potentially
 /// resulting in no affected rows. This envelope provides the necessary context
 /// (e.g. protocol version) to correctly validate the processing of events.
+#[derive(Debug)]
 struct PreparedEventEnvelope {
     metadata: EventMetadata,
     event:    PreparedEvent,
@@ -1712,6 +1765,7 @@ impl PreparedEventEnvelope {
 }
 
 /// Prepared database insertion of a new account.
+#[derive(Debug)]
 struct PreparedAccountCreation {
     /// The base58check representation of the canonical account address.
     account_address:   String,
@@ -1760,6 +1814,7 @@ impl PreparedAccountCreation {
 }
 
 /// Represents the events from an account configuring a delegator.
+#[derive(Debug)]
 struct PreparedAccountDelegationEvents {
     /// Update the state of the delegator.
     events: Vec<PreparedAccountDelegationEvent>,
@@ -1779,6 +1834,7 @@ impl PreparedAccountDelegationEvents {
     }
 }
 
+#[derive(Debug)]
 enum PreparedAccountDelegationEvent {
     StakeIncrease {
         account_id: i64,
@@ -2002,12 +2058,12 @@ impl PreparedAccountDelegationEvent {
                 // some cooldown period, allowing delegators to still target the pool after
                 // removal. Since we remove the baker immediate even for older blocks there
                 // might not be a baker to target, so we check for existence as part of the
-                // query.
+                // query, unless the new target is the passive delegation pool.
                 sqlx::query!(
                     "UPDATE accounts
                         SET delegated_target_baker_id = $1
                     WHERE
-                        EXISTS(SELECT TRUE FROM bakers WHERE id = $1)
+                        ($1::BIGINT IS NULL OR EXISTS(SELECT TRUE FROM bakers WHERE id = $1))
                         AND index = $2",
                     *target_id,
                     account_id
@@ -2027,6 +2083,7 @@ impl PreparedAccountDelegationEvent {
 
 /// Represents the event of a baker being removed, resulting in the delegators
 /// targeting the pool are moved to the passive pool.
+#[derive(Debug)]
 struct BakerRemoved {
     /// Move delegators to the passive pool.
     move_delegators: MovePoolDelegatorsToPassivePool,
@@ -2066,6 +2123,7 @@ impl BakerRemoved {
 }
 
 /// Represents the database operation of removing baker from the baker table.
+#[derive(Debug)]
 struct RemoveBaker {
     baker_id: i64,
 }
@@ -2091,6 +2149,7 @@ impl RemoveBaker {
 
 /// Represents the database operation of moving delegators for a pool to the
 /// passive pool.
+#[derive(Debug)]
 struct MovePoolDelegatorsToPassivePool {
     /// Baker ID of the pool to move delegators from.
     baker_id: i64,
@@ -2119,6 +2178,7 @@ impl MovePoolDelegatorsToPassivePool {
 }
 
 /// Represent the events from configuring a baker.
+#[derive(Debug)]
 struct PreparedBakerEvents {
     /// Update the status of the baker.
     events: Vec<PreparedBakerEvent>,
@@ -2132,13 +2192,17 @@ impl PreparedBakerEvents {
         protocol_version: ProtocolVersion,
     ) -> anyhow::Result<()> {
         for event in &self.events {
-            event.save(tx, transaction_index, protocol_version).await?;
+            event
+                .save(tx, transaction_index, protocol_version)
+                .await
+                .with_context(|| format!("Failed processing baker event {:?}", event))?;
         }
         Ok(())
     }
 }
 
 /// Event changing state related to validators (bakers).
+#[derive(Debug)]
 enum PreparedBakerEvent {
     Add {
         baker_id:             i64,
@@ -2537,6 +2601,7 @@ impl PreparedBakerEvent {
     }
 }
 
+#[derive(Debug)]
 struct PreparedModuleDeployed {
     module_reference: String,
     schema:           Option<Vec<u8>>,
@@ -2583,11 +2648,13 @@ impl PreparedModuleDeployed {
             self.schema
         )
         .execute(tx.as_mut())
-        .await?;
+        .await
+        .with_context(|| format!("Failed inserting into smart_contract_modules: {:?}", self))?;
         Ok(())
     }
 }
 
+#[derive(Debug)]
 struct PreparedModuleLinkAction {
     module_reference:   String,
     contract_index:     i64,
@@ -2638,6 +2705,7 @@ impl PreparedModuleLinkAction {
     }
 }
 
+#[derive(Debug)]
 struct PreparedContractInitialized {
     index:                i64,
     sub_index:            i64,
@@ -2759,7 +2827,7 @@ impl PreparedContractInitialized {
         )
         .execute(tx.as_mut())
         .await
-        .context("Failed inserting new to 'contracts' table")?;
+        .with_context(|| format!("Failed inserting new to 'contracts' table: {:?}", self))?;
 
         self.module_link_event
             .save(tx, transaction_index)
@@ -2777,6 +2845,7 @@ impl PreparedContractInitialized {
 }
 
 /// Represents updates related to rejected transactions.
+#[derive(Debug)]
 enum PreparedRejectedEvent {
     /// Rejected transaction attempting to initialize a smart contract
     /// instance or redeploying a module reference.
@@ -2806,6 +2875,7 @@ impl PreparedRejectedEvent {
     }
 }
 
+#[derive(Debug)]
 struct PreparedRejectModuleTransaction {
     module_reference: String,
 }
@@ -2842,6 +2912,7 @@ impl PreparedRejectModuleTransaction {
     }
 }
 
+#[derive(Debug)]
 struct PreparedContractUpdates {
     /// Additional events to track from the trace elements in the update
     /// transaction.
@@ -2877,12 +2948,19 @@ impl PreparedContractUpdates {
         transaction_index: i64,
     ) -> anyhow::Result<()> {
         for elm in &self.trace_elements {
-            elm.save(tx, transaction_index).await?;
+            elm.save(tx, transaction_index).await.with_context(|| {
+                format!(
+                    "Failed processing contract update trace element with index {} related to \
+                     contract: <{},{}>",
+                    elm.trace_element_index, elm.contract_index, elm.contract_sub_index
+                )
+            })?;
         }
         Ok(())
     }
 }
 
+#[derive(Debug)]
 struct PreparedTraceElement {
     height:              i64,
     contract_index:      i64,
@@ -3049,7 +3127,10 @@ impl PreparedTraceElement {
         .execute(tx.as_mut())
         .await?;
 
-        self.trace_event.save(tx, transaction_index).await?;
+        self.trace_event
+            .save(tx, transaction_index)
+            .await
+            .context("Failed processing trace event")?;
 
         for log in self.cis2_token_events.iter() {
             process_cis2_token_event(
@@ -3059,12 +3140,14 @@ impl PreparedTraceElement {
                 transaction_index,
                 tx,
             )
-            .await?
+            .await
+            .context("Failed processing CIS-2 token event")?
         }
         Ok(())
     }
 }
 
+#[derive(Debug)]
 enum PreparedContractTraceEvent {
     /// Potential module link events from a smart contract upgrade
     Upgrade(PreparedTraceEventUpgrade),
@@ -3083,14 +3166,24 @@ impl PreparedContractTraceEvent {
         transaction_index: i64,
     ) -> anyhow::Result<()> {
         match self {
-            PreparedContractTraceEvent::Upgrade(event) => event.save(tx, transaction_index).await,
-            PreparedContractTraceEvent::Transfer(event) => event.save(tx, transaction_index).await,
-            PreparedContractTraceEvent::Update(event) => event.save(tx, transaction_index).await,
+            PreparedContractTraceEvent::Upgrade(event) => event
+                .save(tx, transaction_index)
+                .await
+                .context("Failed processing contract upgrade trace event"),
+            PreparedContractTraceEvent::Transfer(event) => event
+                .save(tx, transaction_index)
+                .await
+                .context("Failed processing contract transfer of CCD trace even"),
+            PreparedContractTraceEvent::Update(event) => event
+                .save(tx, transaction_index)
+                .await
+                .context("Failed processing contract update trace event"),
             PreparedContractTraceEvent::NoEvent => Ok(()),
         }
     }
 }
 
+#[derive(Debug)]
 struct PreparedTraceEventUpgrade {
     module_removed:        PreparedModuleLinkAction,
     module_added:          PreparedModuleLinkAction,
@@ -3129,6 +3222,7 @@ impl PreparedTraceEventUpgrade {
     }
 }
 
+#[derive(Debug)]
 struct PreparedUpdateContractLastUpgrade {
     contract_index:     i64,
     contract_sub_index: i64,
@@ -3163,6 +3257,7 @@ impl PreparedUpdateContractLastUpgrade {
 }
 
 /// Represent a transfer from contract to an account.
+#[derive(Debug)]
 struct PreparedTraceEventTransfer {
     /// Update the contract balance with the transferred CCD.
     update_contract_balance:  PreparedUpdateContractBalance,
@@ -3203,6 +3298,7 @@ impl PreparedTraceEventTransfer {
     }
 }
 
+#[derive(Debug)]
 struct PreparedTraceEventUpdate {
     /// Update the caller balance (either an account or contract).
     sender:             PreparedTraceEventUpdateSender,
@@ -3210,6 +3306,7 @@ struct PreparedTraceEventUpdate {
     receiving_contract: PreparedUpdateContractBalance,
 }
 
+#[derive(Debug)]
 enum PreparedTraceEventUpdateSender {
     Account(PreparedUpdateAccountBalance),
     Contract(PreparedUpdateContractBalance),
@@ -3249,17 +3346,25 @@ impl PreparedTraceEventUpdate {
         transaction_index: i64,
     ) -> anyhow::Result<()> {
         match &self.sender {
-            PreparedTraceEventUpdateSender::Account(sender) => {
-                sender.save(tx, Some(transaction_index)).await?
-            }
-            PreparedTraceEventUpdateSender::Contract(sender) => sender.save(tx).await?,
+            PreparedTraceEventUpdateSender::Account(sender) => sender
+                .save(tx, Some(transaction_index))
+                .await
+                .context("Failed updating account balance with sending of CCD")?,
+            PreparedTraceEventUpdateSender::Contract(sender) => sender
+                .save(tx)
+                .await
+                .context("Failed updating contract balance with sending of CCD")?,
         }
-        self.receiving_contract.save(tx).await?;
+        self.receiving_contract
+            .save(tx)
+            .await
+            .context("Failed updating contract balance with receiving of CCD")?;
         Ok(())
     }
 }
 
 /// Update of the balance of a contract
+#[derive(Debug)]
 struct PreparedUpdateContractBalance {
     contract_index:     i64,
     contract_sub_index: i64,
@@ -3296,6 +3401,7 @@ impl PreparedUpdateContractBalance {
     }
 }
 
+#[derive(Debug)]
 struct PreparedRejectContractUpdateTransaction {
     contract_index:     i64,
     contract_sub_index: i64,
@@ -3665,8 +3771,14 @@ async fn process_cis2_token_event(
                 serde_json::to_value(cis2_transfer_event)?,
             )
             .execute(tx.as_mut())
-            .await?
-            .ensure_affected_one_row()?;
+            .await
+            .with_context(|| {
+                format!("Failed inserting the token transfer event: {:?}", cis2_transfer_event)
+            })?
+            .ensure_affected_one_row()
+            .with_context(|| {
+                format!("Failed inserting the token transfer event: {:?}", cis2_transfer_event)
+            })?;
         }
 
         // - The `metadata_url` of a token is inserted/updated in the database here.
@@ -3745,6 +3857,7 @@ async fn process_cis2_token_event(
     Ok(())
 }
 
+#[derive(Debug)]
 struct PreparedScheduledReleases {
     canonical_address: CanonicalAccountAddress,
     release_times: Vec<DateTime<Utc>>,
@@ -3825,6 +3938,7 @@ impl PreparedScheduledReleases {
 }
 
 /// Represents either moving funds from or to the encrypted balance.
+#[derive(Debug)]
 struct PreparedUpdateEncryptedBalance {
     /// Update the public balance with the amount being moved.
     public_balance_change: PreparedUpdateAccountBalance,
@@ -3861,6 +3975,7 @@ impl PreparedUpdateEncryptedBalance {
 }
 
 /// Represents change in the balance of some account.
+#[derive(Debug)]
 struct PreparedUpdateAccountBalance {
     /// Address of the account.
     canonical_address: CanonicalAccountAddress,
@@ -3906,8 +4021,20 @@ impl PreparedUpdateAccountBalance {
             self.canonical_address.0.as_slice(),
         )
         .execute(tx.as_mut())
-        .await?
-        .ensure_affected_one_row()?;
+        .await
+        .with_context(|| {
+            format!(
+                "Failed processing update to account balance, change: {}, canonical address: {:?}",
+                self.change, self.canonical_address
+            )
+        })?
+        .ensure_affected_one_row()
+        .with_context(|| {
+            format!(
+                "Failed processing update to account balance, change: {}, canonical address: {:?}",
+                self.change, self.canonical_address
+            )
+        })?;
         // Add the account statement, note that this operation assumes the account
         // balance is already updated.
         self.account_statement.save(tx, transaction_index).await?;
@@ -3916,6 +4043,7 @@ impl PreparedUpdateAccountBalance {
 }
 
 /// Represent the event of a transfer of CCD from one account to another.
+#[derive(Debug)]
 struct PreparedCcdTransferEvent {
     /// Updating the sender account balance.
     update_sender:   PreparedUpdateAccountBalance,
@@ -3954,8 +4082,14 @@ impl PreparedCcdTransferEvent {
         tx: &mut sqlx::Transaction<'static, sqlx::Postgres>,
         transaction_index: i64,
     ) -> anyhow::Result<()> {
-        self.update_sender.save(tx, Some(transaction_index)).await?;
-        self.update_receiver.save(tx, Some(transaction_index)).await?;
+        self.update_sender
+            .save(tx, Some(transaction_index))
+            .await
+            .context("Failed processing sender balance update")?;
+        self.update_receiver
+            .save(tx, Some(transaction_index))
+            .await
+            .context("Failed processing receiver balance update")?;
         Ok(())
     }
 }
@@ -3978,6 +4112,9 @@ impl PreparedSpecialTransactionOutcomes {
         block_info: &BlockInfo,
         events: &[SpecialTransactionOutcome],
     ) -> anyhow::Result<Self> {
+        // Return whether the block is a payday block. This is always false for
+        // protocol versions before P4. In protocol version 4 and later this is the
+        // block where all the rewards are paid out.
         let is_payday_block = events.iter().any(|ev| {
             matches!(
                 ev,
@@ -4490,13 +4627,13 @@ struct PreparedValidatorSuspension {
 /// Represents a payday block, its payday commission
 /// rates, and the associated block height.
 struct PreparedPayDayBlock {
-    block_height:            i64,
+    block_height:                 i64,
     /// Represents the payday baker pool commission rates captured from
     /// the `get_bakers_reward_period` node endpoint.
-    payday_commission_rates: PreparedPaydayCommissionRates,
+    payday_commission_rates:      PreparedPaydayCommissionRates,
     /// Represents the payday lottery power updates for bakers captured from
     /// the `get_election_info` node endpoint.
-    bakers_lottery_powers:   PreparedPaydayLotteryPowers,
+    payday_bakers_lottery_powers: PreparedPaydayLotteryPowers,
 }
 
 impl PreparedPayDayBlock {
@@ -4526,12 +4663,13 @@ impl PreparedPayDayBlock {
             .get_election_info(BlockIdentifier::AbsoluteHeight(block_height))
             .await?
             .response;
-        let bakers_lottery_powers = PreparedPaydayLotteryPowers::prepare(election_info.bakers)?;
+        let payday_bakers_lottery_powers =
+            PreparedPaydayLotteryPowers::prepare(election_info.bakers)?;
 
         Ok(Self {
             block_height: block_height.height.try_into()?,
             payday_commission_rates,
-            bakers_lottery_powers,
+            payday_bakers_lottery_powers,
         })
     }
 
@@ -4543,7 +4681,7 @@ impl PreparedPayDayBlock {
         self.payday_commission_rates.save(tx).await?;
 
         // Save the lottery_powers to the database.
-        self.bakers_lottery_powers.save(tx).await?;
+        self.payday_bakers_lottery_powers.save(tx).await?;
 
         sqlx::query!(
             "UPDATE current_chain_parameters
@@ -4635,15 +4773,23 @@ impl PreparedPaydayCommissionRates {
 struct PreparedPaydayLotteryPowers {
     baker_ids:             Vec<i64>,
     bakers_lottery_powers: Vec<BigDecimal>,
+    ranks:                 Vec<i64>,
 }
 
 impl PreparedPaydayLotteryPowers {
-    fn prepare(bakers: Vec<BirkBaker>) -> anyhow::Result<Self> {
+    fn prepare(mut bakers: Vec<BirkBaker>) -> anyhow::Result<Self> {
         let capacity = bakers.len();
         let mut baker_ids: Vec<i64> = Vec::with_capacity(capacity);
         let mut bakers_lottery_powers: Vec<BigDecimal> = Vec::with_capacity(capacity);
+        let mut ranks: Vec<i64> = Vec::with_capacity(capacity);
 
-        for baker in bakers.iter() {
+        // Sort bakers by lottery power. The baker with the highest lottery power comes
+        // first in the vector and gets rank 1.
+        bakers.sort_by(|self_baker, other_baker| {
+            self_baker.baker_lottery_power.total_cmp(&other_baker.baker_lottery_power).reverse()
+        });
+
+        for (rank, baker) in bakers.iter().enumerate() {
             baker_ids.push(i64::try_from(baker.baker_id.id.index)?);
             bakers_lottery_powers.push(
                 BigDecimal::from_f64(baker.baker_lottery_power)
@@ -4653,11 +4799,13 @@ impl PreparedPaydayLotteryPowers {
                     )
                     .map_err(RPCError::ParseError)?,
             );
+            ranks.push((rank + 1) as i64);
         }
 
         Ok(Self {
             baker_ids,
             bakers_lottery_powers,
+            ranks,
         })
     }
 
@@ -4675,13 +4823,16 @@ impl PreparedPaydayLotteryPowers {
         sqlx::query!(
             "INSERT INTO bakers_payday_lottery_powers (
                 id,
-                payday_lottery_power
+                payday_lottery_power,
+                payday_ranking_by_lottery_powers
             )
             SELECT
                 UNNEST($1::BIGINT[]) AS id,
-                UNNEST($2::NUMERIC[]) AS payday_lottery_power",
+                UNNEST($2::NUMERIC[]) AS payday_lottery_power,
+                UNNEST($3::BIGINT[]) AS payday_ranking_by_lottery_powers",
             &self.baker_ids,
-            &self.bakers_lottery_powers
+            &self.bakers_lottery_powers,
+            &self.ranks
         )
         .execute(tx.as_mut())
         .await?;